Print errors instead of crashing
Interestingly, I got an OverlongHeaders error! That’s very unexpected: I
wonder what caused it? Unfortunately I don’t get to see the headers when
that happened, I’d be very interested to know what they were. I suspect
this is a glitch in their software rather than anything malicious.

Anyway, why was I crashing on unknown errors? The obvious thing to do
seems to be to log it (by printing it) and then continue.