Try flushing log handle more explicitly
The logs weren’t appearing until the app finished, despite the docs
saying that ~withLogTextFile~ flushes the file automatically. This will
hopefully mean that the logs are visible in the file as soon as they
happen, which is what I would like so I can see what’s happening with
the data collection.