Set optimisation differently
From https://github.com/haskell/cabal/issues/6221, it seems like this is
the preferred way to set the optimisation level. I’m also not certain
whether this is taken into account when building with nix: it’s not
mentioned in the output of cabal2nix, so how does the builder know what
optimisation level to use?