Use more and fewer SQL transactions
More as in more of the SQL statements are covered, fewer as in fewer
transactions covering more statements.

I remembered that transactions have another use: ensuring that the code
has a consistent view if it selects and then updates. E.g. if you have
two threads updating a counter by reading and then writing, the read and
write need to happen atomically to avoid missing any updates. So I’ve
aimed to group each read and dependent write(s) within a transaction.